Displays dialog box for when no images are in folder or no matching hashes are found
When no images are found in folder then a dialog box saying "Could not find any images in folder" is displayed with the option to close the program. When no matching images(hashes) are found in folder then a dialog box saying "Could not find any identical images in folder" is displayed with the option to close the program.